feat: Add Wipe Resume Functionality
Added the option to wipe all the resumes on the user homepage. But, currently there is no button for wiping the data from the homepage. So, the user needs to go the <url>/wipe to get access to the wipe button. In the next version, a dedicated wipe button will be added on the homepage.